Fix next block issue; use int16_t for block local index; remove -Wno-sign-compare

+++ src/idx.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <bits/stdint-uintn.h>
-
 #include <cstdint>
 
 namespace ulayfs {
@@ -12,10 +10,10 @@ using LogicalBlockIdx = uint32_t;
 using VirtualBlockIdx = uint32_t;
 
 // local index within a block; this can be -1 to indicate an error
-using BitmapLocalIdx = int32_t;
-using TxLocalIdx = int32_t;
-// Note: LogLocalIdx will persist and we need range [0, 256)
-using LogLocalIdx = uint8_t;
+using BitmapLocalIdx = int16_t;
+using TxLocalIdx = int16_t;
+// Note: LogLocalIdx will persist and the valid range is [0, 255]
+using LogLocalIdx = uint16_t;
 
 // identifier of bitmap blocks; checkout BitmapBlock's doc to see more
 using BitmapBlockId = uint32_t;
